requirements.txt
wget
pyaml